+++ Cargo.toml
@@ -47,4 +47,7 @@ futures = "0.3.12"
 log = "0.4"
 env_logger = "0.9.0"
 pyo3 = { version = "0.16", features = ["extension-module", "abi3-py37"] }
-pyo3-asyncio = { version = "0.16", features = ["attributes", "async-std-runtime"] }
+pyo3-asyncio = { version = "0.16", features = [
+    "attributes",
+    "async-std-runtime",
+] }
